backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9

# PDF processing
pypdf==6.1.3
pdfplumber==0.10.3
pypdfium2==4.24.0
Pillow==10.3.0
pdf2image==1.16.3
pymupdf==1.26.6

# Table extraction
camelot-py[cv]==1.0.9
tabula-py==2.10.0

# Excel export
openpyxl==3.1.5

# Image processing
numpy==1.26.2
opencv-python==4.8.1.78
scipy==1.11.4

# OCR
pytesseract==0.3.10

# Metadata APIs
habanero==1.2.6
biopython==1.81
requests==2.32.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Environment variables
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
